simplify parsing of class to id mapping and move classes to cosem package
