+++ src/main/java/org/janelia/saalfeldlab/cosem/ExtractLabels.java
@@ -1,4 +1,4 @@
-package org.janelia.saalfeldlab;
+package org.janelia.saalfeldlab.cosem;
 
 import net.imglib2.*;
 import net.imglib2.converter.Converters;
@@ -16,9 +16,11 @@ import org.janelia.saalfeldlab.n5.imglib2.N5Utils;
 import picocli.CommandLine;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Paths;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.OptionalDouble;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
@@ -30,46 +32,6 @@ import java.util.concurrent.Executors;
 public class ExtractLabels implements Callable<Void> {
 
     private static int defaultClassId = 100;
-
-    private static class CosemClassToId {
-
-        private final Map<String, Integer> mapping;
-
-        CosemClassToId() throws IOException {
-
-            mapping = new HashMap<>();
-            try (final InputStream stream = getClass().getClassLoader().getResourceAsStream("cosem-classes.txt")) {
-                try (final Scanner scanner = new Scanner(stream)) {
-
-                    while (scanner.hasNext()) {
-                        final int id = scanner.nextInt();
-                        String line = scanner.nextLine().trim();
-                        if (line.indexOf('=') != -1)
-                            line = line.substring(0, line.indexOf('=')).trim();
-
-                        while (line.startsWith("("))
-                            line = line.substring(1).trim();
-                        while (line.endsWith(")"))
-                            line = line.substring(0, line.length() - 1).trim();
-
-                        mapping.put(getClassKey(line), id);
-                    }
-                }
-            }
-        }
-
-        int getClassId(final String name) {
-
-            final String key = getClassKey(name);
-            return mapping.containsKey(key) ? mapping.get(key) : -1;
-        }
-
-        String getClassKey(final String name) {
-
-            return name.trim().replace(' ', '_').toLowerCase();
-        }
-    }
-
 
     public static <T extends NativeType<T> & RealType<T>> void extractLabels(
             final String containerPath,
