+++ src/main/java/org/janelia/saalfeldlab/cosem/ViewCosem.java
@@ -117,7 +117,7 @@
  *
  *
  */
-package org.janelia.saalfeldlab;
+package org.janelia.saalfeldlab.cosem;
 
 import bdv.tools.brightness.ConverterSetup;
 import bdv.tools.brightness.MinMaxGroup;
@@ -140,6 +140,7 @@ import net.imglib2.type.volatiles.AbstractVolatileNativeRealType;
 import net.imglib2.type.volatiles.VolatileDoubleType;
 import net.imglib2.util.Pair;
 import net.imglib2.util.ValuePair;
+import org.janelia.saalfeldlab.N5Factory;
 import org.janelia.saalfeldlab.N5Factory.N5Options;
 import org.janelia.saalfeldlab.n5.N5FSReader;
 import org.janelia.saalfeldlab.n5.N5Reader;
